signal: When the process receives a termination signal, it first terminates mcexec. refs #863 refs #870
original commit-id: 4bb65494e9fd3c9f838409c9d2b29d2b72133077

+++ kernel/include/process.h
@@ -493,7 +493,7 @@ struct process {
 			// V       +----   |
 			// PS_STOPPED -----+
 			// (PS_TRACED)
-	int exit_status; // only for zombie
+	unsigned long exit_status; // only for zombie
 
 	/* Store exit_status for a group of threads when stopped by SIGSTOP.
 	   exit_status can't be used because values of exit_status of threads
